Separate RSC streaming tests into their own file.
